refactor: view, serializer 도메인 분리(#165)
