+++ products/views/__init__.py
@@ -1,0 +1,2 @@
+from .product import ProductDetailView, ProductRankingView, ProductRankingCategoryView, ProductListView, ProductSearchView, MainView, UploadProductImageView, ProductRequestView
+from .category import ProductCategoryView
